backend/requirements.txt
# requirements.txt
Flask>=2.3.3
flask-cors>=4.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
reportlab>=4.0.4
Pillow>=10.0.0
python-magic>=0.4.27
bcrypt>=4.0.1
PyJWT>=2.8.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.2
Jinja2>=3.1.2
gunicorn>=21.2.0
pytest>=7.4.2